Revert "Add integration tests for applications"
This reverts commit ca459f8d86ece4dc1cac21e90b8d808eef8c8720.
